system/outputs/v4/chakin/requirements.txt
chakin
pandas
progressbar2
six
requests
unittest